Comments quantity is back, renaming
