+++ src/components/film-details.js
@@ -192,7 +192,7 @@ const createFilmDetailsTemplate = (film, comments) => {
 
         <div class="form-details__bottom-container">
           <section class="film-details__comments-wrap">
-            <h3 class="film-details__comments-title">Comments <span class="film-details__comments-count"></span></h3>
+            <h3 class="film-details__comments-title">Comments <span class="film-details__comments-count">${comments.length > 0 ? comments.length : ``}</span></h3>
 
             <ul class="film-details__comments-list">
               ${createCommentsTemplate(comments)}
